refactor(dashboard): split governance-panels and swarm components
Extract governance-panels.ts (310 lines) into three focused files:
- governance-panels.ts (141): guardrail pane + action request
- governance-detail.ts (102): decision detail, petition/brief entries
- governance-strips.ts (96): activity rail, freshness strip, runtime params

Extract swarm.ts (293 lines) into two files:
- swarm.ts (98): swarm panel + re-exports
- swarm-surface.ts (203): live-run surface with all section panels

Also fix pre-existing setTimeout type mismatch in command/index.ts.

Closes #1518
