+++ dashboard/src/components/command/swarm.ts
@@ -1,16 +1,5 @@
-import { html } from 'htm/preact'
-import { SwarmOverviewPanel } from './swarm-overview-panel'
-import { SwarmLivePanels } from './swarm-live-panels'
-
-// Re-export for consumers that import from './swarm'
+export { SwarmOverviewPanel } from './swarm-overview-panel'
+export { SwarmLivePanels } from './swarm-live-panels'
 export { SwarmBlockerCard, SwarmChecklistCard, SwarmWorkerCard } from './swarm-cards'
 export { SwarmHealthBar, SwarmRunResolutionCard, SwarmStoryboard } from './swarm-storyboard'
-
-export function SwarmSurface() {
-  return html`
-    <div class="command-section-stack">
-      <${SwarmOverviewPanel} />
-      <${SwarmLivePanels} />
-    </div>
-  `
-}
+export { SwarmSurface } from './swarm-surface'
